[update] 전시회 목록 페이지 컴포넌트 분리  #17
